fix(/Repository)[ResultFileManager.cs]: handler multiple lib adding
When you want to add libs for specific command, for now the `echo`
command, we should have one lib instead of multiple libs which added for
each command. So, I added a list which will remember the libs that i
used. Now on if any process calls the lib add method, it will first
check that the lib was added ealer or not and then it will add the
specific lib.